Reorganization fixes
This commit brings some reorganization fixes like:

	tortoise_installer -> /etc/tortoise
	install.sh -> tortoise_installer
	egginstall -> symlinks to -> tortoise_installer/install.sh

More changes:

- now you don't need password to login.
- also dont need password to use sudo.
- lightdm was removed from liveISO

for future commits

more reorganization fixes, improvements on code and
revisions on configurations or installation, but
no new features for now.

I'm focused on creating a solid environment so that
I can work better on the other details in the future.
Think of it as me working on the foundation of all of this.